Share ASR evaluation helpers
