+++ asr/result_io.py
@@ -1,0 +1,13 @@
+"""Small helpers for writing evaluation results safely."""
+
+import json
+from pathlib import Path
+
+
+def write_json(path: Path, data: dict) -> None:
+    path.parent.mkdir(parents=True, exist_ok=True)
+    temporary = path.with_suffix(path.suffix + ".tmp")
+    temporary.write_text(
+        json.dumps(data, ensure_ascii=False, indent=2), encoding="utf-8"
+    )
+    temporary.replace(path)
